Crate range_map

Structs

EltIter

PairIter

Range
A range of elements, including the endpoints.

RangeIter

RangeMap
A set of characters. Optionally, each character in the set may be associated with some data.

RangeMultiMap
A multi-valued mapping from primitive integers to other data.

RangeSet
A set of integers, implemented as a sorted list of (inclusive) ranges.