Struct range_map::RangeMap

pub struct RangeMap<T> {
    // some fields omitted
}

A RangeMap stores values of T that map to Ranges.

Methods

impl<T> RangeMap<T>

fn new() -> RangeMap<T>

Create a new, empty RangeMap.

fn from(entries: Vec<Entry<T>>) -> RangeMap<T>

Create a RangeMap with entries.

fn len(&self) -> usize

Returns the number of entries in the RangeMap.

fn insert(&mut self, range: Range, value: T) -> Result<(), ()>

Insert value in range.

fn lookup(&self, addr: Addr) -> Option<&T>

Find an entry whose Range encompasses addr.

fn iter(&self) -> Iter<Entry<T>>

Return an iterator over the entries of the RangeMap.