Struct range_map::RangeMap

pub struct RangeMap<T> {
    // some fields omitted
}

A RangeMap stores values of T that map to Ranges.

Methods

impl<T> RangeMap<T>

fn new() -> RangeMap<T>

Create a new, empty RangeMap.

fn insert(&mut self, (start, end): Range, value: T) -> Result<(), ()>

Insert value in the range (start, end).

fn lookup(&self, addr: Addr) -> Option<&T>

Find an entry whose Range encompasses addr.

fn iter(&self) -> Iter<Entry<T>>

Return an iterator over the entries of the RangeMap.