Struct range_map::RangeMap

pub struct RangeMap<T, V> { /* fields omitted */ }

A set of characters. Optionally, each character in the set may be associated with some data.

Methods

impl<T: Debug + PrimInt, V: Clone + Debug + Eq> RangeMap<T, V>

pub fn new() -> RangeMap<T, V>

Creates a new empty RangeMap.

pub fn from_sorted_vec(vec: Vec<(Range<T>, V)>) -> RangeMap<T, V>

Creates a RangeMap from a Vec, which must contain ranges in ascending order. If any ranges overlap, they must map to the same value.

Panics if the ranges are not sorted, or if they overlap without mapping to the same value.

pub fn num_ranges(&self) -> usize

Returns the number of mapped ranges.

Note that this is not usually the same as the number of mapped values.

pub fn is_empty(&self) -> bool

Tests whether this map is empty.

pub fn is_full(&self) -> bool

Tests whether this CharMap maps every value.

pub fn ranges_values<'a>(&'a self) -> Iter<'a, (Range<T>, V)>

Iterates over all the mapped ranges and values.

Important traits for PairIter<'a, T, V>
impl<'a, T: PrimInt, V> Iterator for PairIter<'a, T, V> type Item = (T, &'a V);

pub fn keys_values<'a>(&'a self) -> PairIter<'a, T, V>

Iterates over all mappings.

pub fn get(&self, x: T) -> Option<&V>

Finds the value that x maps to, if it exists.

Runs in O(log n) time, where n is the number of mapped ranges.

pub fn intersection(&self, other: &RangeSet<T>) -> RangeMap<T, V>

Returns those mappings whose keys belong to the given set.

pub fn num_keys(&self) -> usize

Counts the number of mapped keys.

This saturates at usize::MAX.

pub fn to_range_set(&self) -> RangeSet<T>

Returns the set of mapped chars, forgetting what they are mapped to.

pub fn map_values<F>(&mut self, f: F) where
F: FnMut(&V) -> V,

Modifies the values in place.

pub fn retain_values<F>(&mut self, f: F) where
F: FnMut(&V) -> bool,

Modifies this map to contain only those mappings with values v satisfying f(v).

pub fn as_mut_slice(&mut self) -> &mut [(Range<T>, V)]

Returns a mutable view into this map.

The ranges should not be modified, since that might violate our invariants.

This method will eventually be removed, probably once anonymous return values allow is to write a values_mut() iterator more easily.