Struct range_map::RangeSet

pub struct RangeSet<T> { /* fields omitted */ }

A set of integers, implemented as a sorted list of (inclusive) ranges.

Methods

impl<T: Debug + PrimInt> RangeSet<T>

pub fn new() -> RangeSet<T>

Creates a new empty RangeSet.

pub fn is_empty(&self) -> bool

Tests if this set is empty.

pub fn is_full(&self) -> bool

Tests whether this set contains every valid value of T.

pub fn num_ranges(&self) -> usize

Returns the number of ranges used to represent this set.

pub fn num_elements(&self) -> usize

Returns the number of elements in the set.

This saturates at usize::MAX.

Important traits for RangeIter<'a, T>
impl<'a, T: Debug + PrimInt> Iterator for RangeIter<'a, T> type Item = Range<T>;

pub fn ranges<'a>(&'a self) -> RangeIter<'a, T>

Returns an iterator over all ranges in this set.

Important traits for EltIter<'a, T>
impl<'a, T: Debug + PrimInt> Iterator for EltIter<'a, T> type Item = T;

pub fn elements<'a>(&'a self) -> EltIter<'a, T>

Returns an iterator over all elements in this set.

pub fn contains(&self, val: T) -> bool

Checks if this set contains a value.

pub fn union(&self, other: &RangeSet<T>) -> RangeSet<T>

Returns the union between self and other.

pub fn full() -> RangeSet<T>

Creates a set that contains every value of T.

pub fn single(x: T) -> RangeSet<T>

Creates a set containing a single element.

pub fn except<I: Iterator<Item = T>>(it: I) -> RangeSet<T>

Creates a set containing all elements except the given ones.

Panics

if chars is not sorted or not unique.

pub fn intersection(&self, other: &RangeSet<T>) -> RangeSet<T>

Finds the intersection between this set and other.

pub fn negated(&self) -> RangeSet<T>

Returns the set of all characters that are not in this set.